Split actions/index.js into actions_recipe/activity.js
