+++ src/actions/actions_recipe.jsx
@@ -6,12 +6,6 @@ export const FETCH_RECIPE = "fetch_recipe";
 export const CREATE_RECIPE = "create_recipe";
 export const DELETE_RECIPE = "delete_recipe";
 
-export const FETCH_ACTIVITIES = "fetch_activities";
-export const FETCH_ACTIVITY = "fetch_activity";
-export const CREATE_ACTIVITY = "create_activity";
-export const DELETE_ACTIVITY = "delete_activity";
-
-
 export function fetchRecipes() {
   const request = axios.get(`${ROOT_URL}/recipes/`);
 
@@ -64,57 +58,6 @@ export function deleteRecipe(id, callback) {
 
   return {
     type: DELETE_RECIPE,
-    payload: id
-  };
-}
-
-export function fetchActivities() {
-  const request = axios.get(`${ROOT_URL}/activities/`);
-
-  return {
-    type: FETCH_ACTIVITIES,
-    payload: request
-  };
-}
-
-export function createActivity(values, callback) {
-
-  const request = axios
-    .post(`${ROOT_URL}/activities/`, values, {
-      headers: {
-        Authorization: "Token " + localStorage.getItem("token")
-      }
-    })
-    .then(() => callback());
-
-  return {
-    type: CREATE_ACTIVITY,
-    payload: request
-  };
-}
-
-export function fetchActivity(id) {
-  const request = axios.get(`${ROOT_URL}/activities/${id}/`);
-
-  return {
-    type: FETCH_ACTIVITY,
-    payload: request
-  };
-}
-
-export function deleteActivity(id, callback) {
-
-  // eslint-disable-next-line
-  const request = axios
-    .delete(`${ROOT_URL}/activities/${id}/`, {
-      headers: {
-        Authorization: "Token " + localStorage.getItem("token")
-      }
-    })
-    .then(() => callback());
-
-  return {
-    type: DELETE_ACTIVITY,
     payload: id
   };
 }
